feat: implement SignUp Controller Test
