+++ src/main/java/org/swmaestro/repl/gifthub/auth/controller/AuthController.java
@@ -1,0 +1,4 @@
+package org.swmaestro.repl.gifthub.auth.controller;
+
+public class AuthController {
+}
