+++ src/main/java/org/swmaestro/repl/gifthub/auth/repository/SpringDataJpaMemberRepository.java
@@ -1,7 +1,7 @@
-package org.swmaestro.repl.gifthub.member.repository;
+package org.swmaestro.repl.gifthub.auth.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.swmaestro.repl.gifthub.member.entity.Member;
+import org.swmaestro.repl.gifthub.auth.entity.Member;
 
 public interface SpringDataJpaMemberRepository extends JpaRepository<Member, Long> {
 	Member findByUsername(String username);
